requirements.txt
angler==0.0.15
autograd==1.6.2
h5py==3.11.0
ipython==8.25.0
matplotlib==3.8.4
meep==1.0.6
mlflow==2.13.0
mlflow_skinny==2.13.0
nlopt==2.7.1
numpy==2.1.0
opencv_python==4.8.1.78
opencv_python_headless==4.10.0.84
pyMKL==0.0.3
pyMKL==0.0.3
pypardiso==0.4.6
scipy==1.14.1
torch==2.3.0
torchonn_pyutils==0.0.3.1
torchonn_pyutils.egg==info
